Move timeout up to Fog from Fog::Mock.
