+++ lib/fog/core.rb
@@ -31,4 +31,5 @@ require 'fog/core/service'
 require 'fog/core/ssh'
 require 'fog/core/scp'
 require 'fog/core/time'
+require 'fog/core/timeout'
 require 'fog/core/wait_for'
